refactor: reorganize source files
